requirements.txt
flask
google-generativeai
pandas
beautifulsoup4
requests
urllib3
